pyproject.toml
[tool.poetry]
name = "events-backend"
version = "0.1.0"
description = ""
authors = ["ivankryukov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
